Seperate Legend in 2 components
